+++ config/config.exs
@@ -6,5 +6,6 @@ config :td_core, TdCore.Search.Cluster,
   aliases: %{test_alias: "string_test_alias"}
 
 config :td_core, TdCore.Search.IndexWorker, TdCore.Search.IndexWorkerMock
+config :td_cluster, TdCluster.ClusterHandler, MockClusterHandler
 
 config :td_cache, redis_host: "redis", port: 6380